Form endpoints updated
- Se actualiza index.js, a modo de modularizar los endpoints.
- Se crea form.js, que contiene las funciones .post y .get, referidas a form.js.
